+++ packages/api-reference/src/components/Content/Introduction/Introduction.vue
@@ -5,9 +5,9 @@ import GitHubSlugger from 'github-slugger'
 import { computed, onMounted } from 'vue'
 
 import { SpecificationExtension } from '@/components/SpecificationExtension'
-import { DEFAULT_INTRODUCTION_SLUG } from '@/hooks'
 import { useConfig } from '@/hooks/useConfig'
 import { useNavState } from '@/hooks/useNavState'
+import { DEFAULT_INTRODUCTION_SLUG } from '@/hooks/useSidebar'
 
 import DownloadLink from '../../../features/DownloadLink/DownloadLink.vue'
 import { Badge } from '../../Badge'
@@ -88,9 +88,9 @@ onMounted(() => config.value.onLoaded?.())
             {{ info.title }}
           </SectionHeaderTag>
         </SectionHeader>
-        <DownloadLink :specTitle="filenameFromTitle" />
         <SectionColumns>
           <SectionColumn>
+            <DownloadLink :filename="filenameFromTitle" />
             <Description :value="info.description" />
           </SectionColumn>
           <SectionColumn v-if="$slots.aside">