+++ packages/api-reference/src/components/Content/Schema/SchemaPropertyHeading.vue
@@ -1,4 +1,5 @@
 <script lang="ts" setup>
+import { isDefined } from '@scalar/oas-utils/helpers'
 import type {
   OpenAPI,
   OpenAPIV2,
@@ -6,8 +7,12 @@ import type {
   OpenAPIV3_1,
 } from '@scalar/openapi-types'
 import { stringify } from 'flatted'
+import { computed } from 'vue'
 
-import { discriminators } from '@/components/Content/Schema/helpers/optimizeValueForDisplay'
+import {
+  compositions,
+  type CompositionKeyword,
+} from '@/components/Content/Schema/helpers/schema-composition'
 import SchemaPropertyExamples from '@/components/Content/Schema/SchemaPropertyExamples.vue'
 import ScreenReader from '@/components/ScreenReader.vue'
 
@@ -33,21 +38,33 @@ const {
     | unknown
 }>()
 
-const discriminatorType = discriminators.find((r) => {
+const composition = compositions.find((composition: CompositionKeyword) => {
   if (!value || typeof value !== 'object') {
     return false
   }
 
   return (
-    r in value ||
-    (value.items && typeof value.items === 'object' && r in value.items)
+    composition in value ||
+    (value.items &&
+      typeof value.items === 'object' &&
+      composition in value.items)
   )
 })
 
 const flattenDefaultValue = (value: Record<string, any>) => {
-  return Array.isArray(value?.default) && value.default.length === 1
-    ? value.default[0]
-    : value?.default
+  if (value?.default === null) {
+    return 'null'
+  }
+
+  if (Array.isArray(value?.default) && value.default.length === 1) {
+    return value.default[0]
+  }
+
+  if (typeof value?.default === 'string') {
+    return JSON.stringify(value.default)
+  }
+
+  return value?.default
 }
 
 // Get model name from schema
@@ -56,20 +73,102 @@ const getModelNameFromSchema = (schema: OpenAPI.Document): string | null => {
     return null
   }
 
+  if (schema.title) {
+    return schema.title
+  }
+
   if (schema.name) {
     return schema.name
   }
 
   if (schemas && typeof schemas === 'object') {
+    // Handle direct schema match
     for (const [schemaName, schemaValue] of Object.entries(schemas)) {
       if (stringify(schemaValue) === stringify(schema)) {
         return schemaName
+      }
+    }
+
+    // Handle case where schema is a reference to a component schema
+    if (schema.type === 'object' && schema.properties) {
+      for (const [schemaName, schemaValue] of Object.entries(schemas)) {
+        if (
+          stringify(schemaValue.properties) === stringify(schema.properties)
+        ) {
+          return schemaName
+        }
       }
     }
   }
 
   return null
 }
+
+/** Get the const value from the schema */
+const constValue = computed(() => {
+  if (isDefined(value?.const)) {
+    return value?.const
+  }
+
+  if (value?.enum?.length === 1) {
+    return value.enum[0]
+  }
+
+  if (value?.items) {
+    if (isDefined(value.items.const)) {
+      return value.items.const
+    }
+
+    if (value.items.enum?.length === 1) {
+      return value.items.enum[0]
+    }
+  }
+  return null
+})
+
+/** Computes the human-readable type for a schema property. */
+const displayType = computed(() => {
+  if (Array.isArray(value?.type)) {
+    return value.type.join(' | ')
+  }
+
+  if (value?.title) {
+    return value.title
+  }
+
+  if (value?.name) {
+    return value.name
+  }
+
+  return value?.type ?? ''
+})
+
+/** Format the type and model name */
+const formatTypeWithModel = (type: string, modelName: string) => {
+  return type === 'array' ? `${type} ${modelName}[]` : `${type} ${modelName}`
+}
+
+/** Gets the model name */
+const modelName = computed(() => {
+  if (!value?.type) {
+    return null
+  }
+
+  // Handle array types with item references
+  if (value.type === 'array' && value.items?.type) {
+    const itemModelName =
+      getModelNameFromSchema(value.items) || value.items.type
+    return formatTypeWithModel(value.type, itemModelName)
+  }
+
+  // Handle direct object references
+  const objectModelName = getModelNameFromSchema(value)
+  if (objectModelName) {
+    return objectModelName
+  }
+
+  return null
+})
 </script>
 <template>
   <div class="property-heading">
@@ -83,45 +182,23 @@ const getModelNameFromSchema = (schema: OpenAPI.Document): string | null => {
       <template v-else>&sol;<slot name="name" />&sol;</template>
     </div>
     <div
-      v-if="additional"
-      class="property-additional">
-      <template v-if="value?.['x-additionalPropertiesName']">
-        {{ value['x-additionalPropertiesName'] }}
-      </template>
-      <template v-else>additional properties</template>
+      v-if="value?.isDiscriminator"
+      class="property-discriminator">
+      Discriminator
     </div>
-    <div
-      v-if="pattern"
-      class="property-pattern">
-      <Badge>pattern</Badge>
-    </div>
-    <div
-      v-if="value?.deprecated"
-      class="property-deprecated">
-      <Badge>deprecated</Badge>
-    </div>
-    <div
-      v-if="value?.const || (value?.enum && value.enum.length === 1)"
-      class="property-const">
-      <SchemaPropertyDetail truncate>
-        <template #prefix>const:</template>
-        {{ value.const ?? value.enum[0] }}
-      </SchemaPropertyDetail>
-    </div>
-    <template v-else-if="value?.type">
-      <SchemaPropertyDetail>
+    <template v-if="value">
+      <SchemaPropertyDetail v-if="value?.type">
         <ScreenReader>Type:</ScreenReader>
-        <template v-if="value?.items?.type">
-          {{ value.type }}
-          {{ getModelNameFromSchema(value.items) || value.items.type }}[]
+        <template v-if="modelName">
+          {{ modelName }}
         </template>
         <template v-else>
-          {{ Array.isArray(value.type) ? value.type.join(' | ') : value.type }}
+          {{ displayType }}
           {{ value?.nullable ? ' | nullable' : '' }}
         </template>
-        <template v-if="value.minItems || value.maxItems">
-          {{ value.minItems }}&hellip;{{ value.maxItems }}
-        </template>
+      </SchemaPropertyDetail>
+      <SchemaPropertyDetail v-if="value.minItems || value.maxItems">
+        {{ value.minItems }}&hellip;{{ value.maxItems }}
       </SchemaPropertyDetail>
       <SchemaPropertyDetail v-if="value.minLength">
         <template #prefix>min:</template>
@@ -173,8 +250,34 @@ const getModelNameFromSchema = (schema: OpenAPI.Document): string | null => {
         {{ flattenDefaultValue(value) }}
       </SchemaPropertyDetail>
     </template>
+    <div
+      v-if="additional"
+      class="property-additional">
+      <template v-if="value?.['x-additionalPropertiesName']">
+        {{ value['x-additionalPropertiesName'] }}
+      </template>
+      <template v-else>additional properties</template>
+    </div>
+    <div
+      v-if="pattern"
+      class="property-pattern">
+      <Badge>pattern</Badge>
+    </div>
+    <div
+      v-if="value?.deprecated"
+      class="property-deprecated">
+      <Badge>deprecated</Badge>
+    </div>
+    <div
+      v-if="isDefined(constValue)"
+      class="property-const">
+      <SchemaPropertyDetail truncate>
+        <template #prefix>const:</template>
+        {{ constValue }}
+      </SchemaPropertyDetail>
+    </div>
     <template v-else>
-      <!-- Shows only when a discriminator is used (so value?.type is undefined) -->
+      <!-- Shows only when a composition is used (so value?.type is undefined) -->
       <SchemaPropertyDetail v-if="value?.nullable === true">
         nullable
       </SchemaPropertyDetail>
@@ -198,8 +301,7 @@ const getModelNameFromSchema = (schema: OpenAPI.Document): string | null => {
     <SchemaPropertyExamples
       v-if="withExamples"
       :examples="value?.examples"
-      :example="value?.example || value?.items?.example"
-      :discriminator-type="discriminatorType" />
+      :example="value?.example || value?.items?.example" />
   </div>
 </template>
 <style scoped>
@@ -232,7 +334,9 @@ const getModelNameFromSchema = (schema: OpenAPI.Document): string | null => {
   font-family: var(--scalar-font-code);
   font-weight: var(--scalar-semibold);
   font-size: var(--scalar-font-size-3);
-  display: flex;
+  overflow: hidden;
+  white-space: normal;
+  overflow-wrap: break-word;
 }
 
 .property-additional {
@@ -258,6 +362,12 @@ const getModelNameFromSchema = (schema: OpenAPI.Document): string | null => {
   font-size: var(--scalar-micro);
   color: var(--scalar-color-green);
 }
+
+.property-discriminator {
+  font-size: var(--scalar-micro);
+  color: var(--scalar-color-purple);
+}
+
 .property-detail {
   font-size: var(--scalar-micro);
   color: var(--scalar-color-2);