+++ packages/api-reference/src/features/ApiClientModal/ApiClientModal.vue
@@ -5,16 +5,17 @@ import { getServersFromOpenApiDocument } from '@scalar/oas-utils/transforms'
 import type { ApiClientConfiguration } from '@scalar/types/api-reference'
 import type { Spec } from '@scalar/types/legacy'
 import { watchDebounced } from '@vueuse/core'
-import { useExampleStore } from '#legacy'
 import microdiff from 'microdiff'
 import { onBeforeUnmount, onMounted, ref, watch } from 'vue'
 
-import { useNavState } from '@/hooks'
+import { useNavState } from '@/hooks/useNavState'
+import { useExampleStore } from '@/legacy/stores'
 
 import { useApiClient } from './useApiClient'
 
 const { configuration, parsedSpec } = defineProps<{
-  configuration: Partial<ApiClientConfiguration>
+  // The plugins for @scalar/api-reference and @scalar/api-client are different (as of now, doesn’t have to be).
+  configuration: Partial<Omit<ApiClientConfiguration, 'plugins'>>
   parsedSpec: Spec
 }>()